profile component + orderwithtotal
