+++ client-side/src/app/views/dashboards/dashboard/dashboard.component.ts
@@ -74,9 +74,7 @@ export class DashboardComponent implements OnInit {
 
   getOrders() {
     this.orderService.getOrders().subscribe(
-      orderList => {
-        this.orders = orderList;
-      }
+      orderList => this.orders = orderList
     );
   }
