+++ client-side/src/app/views/profile/profile.component.scss
@@ -23,3 +23,7 @@
         box-shadow: 0 3px 10px 0 rgba(0, 0, 0, 0.15), 0 3px 12px 0 rgba(0, 0, 0, 0.15);
     }
 }
+
+.user-img {
+    height: 10rem;
+}
