+++ client-side/src/app/views/profile/profile.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-profile',
+  templateUrl: './profile.component.html',
+  styleUrls: ['./profile.component.scss']
+})
+export class ProfileComponent implements OnInit {
+  name = 'Corgeek';
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
