--- client-side/src/app/views/profile/profile/profile.component.html
@@ -1,50 +1,0 @@
-<div class="row">
-  <div class="col-md-12">
-    <!--Projects section v.2-->
-    <section class="section text-center pb-3">
-      <!--Grid row-->
-      <div class="row d-flex justify-content-center">
-        <!--Grid column-->
-        <!--Grid column-->
-
-        <!--Grid column-->
-        <div class="col-lg-12 col-xl-12 mb-3">
-
-          <!--Card-->
-          <div class="card card-cascade narrower d-flex mb-5">
-
-            <!--Card image-->
-            <div class="view overlay hm-white-slight">
-              <img src="https://mdbootstrap.com/img/Mockups/Horizontal/6-col/pro-landing.jpg" class="img-fluid" alt="">
-              <a href="#">
-                <div class="mask"></div>
-              </a>
-            </div>
-
-            <!--Card content-->
-            <div class="card-body">
-              <!--Title-->
-              <h4 class="card-title font-bold mb-3"><strong>Project name</strong></h4>
-              <!--Text-->
-              <p class="card-text">Some quick example text to build on the card title and make up the bulk of the
-                card's content.
-              </p>
-
-            </div>
-
-            <!--Card footer-->
-            <div class="card-footer links-light">
-              <span class="right">
-                <a class="waves-effect p-2">Live Preview <i class="fa fa-photo ml-1"></i></a>
-              </span>
-            </div>
-          </div>
-          <!--/.Card-->
-        </div>
-        <!--Grid column-->
-      </div>
-      <!--Grid row-->
-    </section>
-    <!--/Projects section v.2-->
-  </div>
-</div>
