+++ client-side/src/app/views/views.module.ts
@@ -25,7 +25,7 @@ import { UtilitiesComponent } from './css/utilities/utilities.component';
 import { ImagesComponent } from './css/images/images.component';
 import { ColorsComponent } from './css/colors/colors.component';
 import { ShadowComponent } from './css/shadow/shadow.component';
-import { ProfileComponent } from './profile/profile/profile.component';
+import { ProfileComponent } from './profile/profile.component';
 import { HelpComponent } from './help/help.component';
